+++ src/main/kotlin/org/jetbrains/intellij/tasks/InitializeIntelliJPluginTask.kt
@@ -3,6 +3,7 @@
 package org.jetbrains.intellij.tasks
 
 import com.jetbrains.plugin.structure.base.utils.create
+import com.jetbrains.plugin.structure.base.utils.outputStream
 import org.gradle.api.DefaultTask
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Internal
@@ -14,7 +15,10 @@ import org.jetbrains.intellij.IntelliJPluginConstants.PLUGIN_GROUP_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.PLUGIN_ID
 import org.jetbrains.intellij.IntelliJPluginConstants.PLUGIN_NAME
 import org.jetbrains.intellij.utils.LatestVersionResolver
-import java.io.File
+import java.nio.file.Path
+import java.util.jar.JarOutputStream
+import java.util.jar.Manifest
+import kotlin.io.path.exists
 
 /**
  * Initializes the Gradle IntelliJ Plugin and performs various checks, like if the plugin is up to date.
@@ -29,7 +33,10 @@ abstract class InitializeIntelliJPluginTask : DefaultTask() {
     abstract val selfUpdateCheck: Property<Boolean>
 
     @get:Internal
-    abstract val lockFile: Property<File>
+    abstract val selfUpdateLockPath: Property<Path>
+
+    @get:Internal
+    abstract val coroutinesJavaAgentPath: Property<Path>
 
     init {
         group = PLUGIN_GROUP_NAME
@@ -41,13 +48,14 @@ abstract class InitializeIntelliJPluginTask : DefaultTask() {
     @TaskAction
     fun initialize() {
         checkPluginVersion()
+        createCoroutinesJavaAgentFile()
     }
 
     /**
-     * Checks if the plugin is up to date.
+     * Checks if the plugin is up-to-date.
      */
     private fun checkPluginVersion() {
-        if (!selfUpdateCheck.get() || offline.get()) {
+        if (!selfUpdateCheck.get() || selfUpdateLockPath.get().exists() || offline.get()) {
             return
         }
 
@@ -60,9 +68,30 @@ abstract class InitializeIntelliJPluginTask : DefaultTask() {
                 warn(context, "$PLUGIN_NAME is outdated: $version. Update `$PLUGIN_ID` to: $latestVersion")
             }
 
-            lockFile.get().toPath().create()
+            selfUpdateLockPath.get().create()
         } catch (e: Exception) {
             error(context, e.message.orEmpty(), e)
         }
+    }
+
+    /**
+     * Creates a Java Agent file for the Coroutines library required to enable coroutines debugging.
+     */
+    private fun createCoroutinesJavaAgentFile() {
+        if (coroutinesJavaAgentPath.get().exists()) {
+            return
+        }
+
+        val manifest = Manifest(
+            """
+            Manifest-Version: 1.0
+            Premain-Class: kotlinx.coroutines.debug.AgentPremain
+            Can-Retransform-Classes: true
+            Multi-Release: true
+            
+            """.trimIndent().byteInputStream()
+        )
+
+        JarOutputStream(coroutinesJavaAgentPath.get().outputStream(), manifest).close()
     }
 }