+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Configure all tasks that extend task classes instead of just those created by the plugin 
 - Make JbrResolver prefer Gradle javaToolchains by `JetBrains s.r.o`, if available. Only otherwise start fetching and running a new one.
+- Support for Kotlin Coroutines debugging
 
 ### Changed
 - Disabled caching for `BuildPluginTask`